+++ lib/votd.rb
@@ -5,7 +5,6 @@ require "httparty"
 # This class acts as the entry point to all sub-modules that are used
 # to interact with various Verse of the Day web services.
 module Votd
-  include HTTParty
 end
 
 require "votd/version"